ov006: promote 41 dScMgBomroom_c functions into one TU
Folds 41 of the run's 79 legacy one-function shards into
src/actors/dScMgBomroom_c.cpp -- ROM ordinals 14..54, the contiguous
.text run 0x020d5eb8..0x020d7c4c. The claim is PARTIAL by necessity,
and a licensed .text claim cannot span a hole:

  * the run's LOWER edge is ordinal 13, func_ov006_020d5e5c
    (0x020d5e5c..0x020d5eb8). It has a legacy source but no `complete`
    marker -- a 10-word pure register permutation that no admissible
    spelling reproduces, so dsd serves the cartridge's bytes for it.
  * the run's UPPER edge is ordinal 55, func_ov006_020d7c4c
    (0x020d7c4c, 0x230 bytes), the class range's only genuinely
    SOURCELESS function: no file under src/ and no delinks.txt entry.

Parsed from delinks.txt, the class's linker run 0x020d5a54..0x020d9574
holds 79 delink blocks plus that one sourceless function, with exactly
one coverage gap -- the 0x230 bytes at 0x020d7c4c. The sourceless split
is 55 blocks below it and 24 above; this TU is inside the 55-block side.
Of the four all-matching runs between the barriers -- ordinals 0..12
(13), 14..54 (41), 56..59 (4) and 61..79 (19) -- 41 is the largest.

The 17 queue-flagged pragma collisions reduce to 11 brackets inside the
licensed range. All 11 move bytes; none is inert. They bind only under
`#pragma defer_codegen off`, which also forces ROM-ascending source
order, so the file is written ascending from the outset:

  control C1  file as written                    41/41 MATCH
  control C2  all 11 push/pop brackets deleted   30/41 (11 DIFF)
  control C3  brackets kept, defer_codegen on    30/41 (same 11)
  control C4  pragmas hoisted to file scope      30/41 (same 11)
  11 single-bracket removals                     40/41 each, each
                                                 costing exactly its
                                                 own member

Because the destructor pair (ordinals 0/1) sits on the far side of the
ordinal-13 hole, the key function is outside the licensed range: the TU
emits no _ZTV/_ZTI/_ZTS, compiler_only_output is [], externalized_output
is absent, and include/dScMgBomroom_c.h needs no edit.

tubuild verify   41/41 MATCH, objisolate clean, reloc-destinations clean
tubuild linkcheck 41 LICENSED / 0 COLLIDES-GAP / 0 COLLIDES-SRC /
                 0 HOMELESS; emitted .text ROM-ascending; module
                 fidelity 106/106 exact; ROM sha256 identical to the
                 source-independent stock control; dsd check symbols
                 --fail still FAILS on the same 9 pre-existing errors
                 (0 new, 0 resolved) -> LINK-VERIFIED

include/decl_common.h: func_ov006_020d672c was recorded as `int f()`
from the legacy shard's K&R forward declaration rather than its
definition; the cartridge's bytes take the scene pointer in r0 and
leave r0 untouched on the early-out path, so it is corrected to
void(void*).
