+++ attribution.json
@@ -2203,6 +2203,47 @@
     "src/actors/dScMgSound_c.cpp#func_ov006_0211ae40": "tangosdev",
     "src/actors/dScMgSound_c.cpp#func_ov006_0211aa94": "andrewboudreau",
     "src/actors/dScMgSound_c.cpp#func_ov006_0211a4b0": "andrewboudreau",
-    "src/actors/dScMgSound_c.cpp#func_ov006_02119b00": "andrewboudreau"
+    "src/actors/dScMgSound_c.cpp#func_ov006_02119b00": "andrewboudreau",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d5eb8": "andrew",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d5f28": "tangosdev",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d5f2c": "andrewboudreau",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d5fd8": "tangosdev",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d5fec": "tangosdev",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d604c": "tangosdev",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d6084": "andrewboudreau",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d6098": "andrewboudreau",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d6100": "tangosdev",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d6170": "andrewboudreau",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d61dc": "andrewboudreau",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d6264": "tangosdev",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d6278": "andrewboudreau",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d62e0": "andrewboudreau",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d634c": "tangosdev",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d63ac": "tangosdev",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d63d4": "andrewboudreau",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d6454": "tangosdev",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d64c4": "tangosdev",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d64c8": "andrewboudreau",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d65b4": "tangosdev",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d65c8": "andrewboudreau",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d6630": "andrewboudreau",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d669c": "tangosdev",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d66c4": "noreply",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d672c": "tangosdev",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d6784": "andrewboudreau",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d68a8": "andrewboudreau",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d69b8": "andrewboudreau",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d6b88": "tangosdev",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d6c90": "tangosdev",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d6d7c": "andrewboudreau",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d6e8c": "andrewboudreau",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d7524": "andrewboudreau",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d7604": "andrewboudreau",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d7778": "tangosdev",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d777c": "ruspecial",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d7958": "tangosdev",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d795c": "andrewboudreau",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d7a84": "noreply",
+    "src/actors/dScMgBomroom_c.cpp#func_ov006_020d7c00": "andrewboudreau"
   }
 }
