--- src/func_ov006_020d63ac.c
@@ -1,4 +1,0 @@
-void func_ov006_020d63ac(char (*c)[16]){
-  int i;
-  for(i=0;i<2;i++){ c[i][0x628d]=0; c[i][0x628e]=0; }
-}
