--- src/func_ov006_020d64c4.c
@@ -1,3 +1,0 @@
-void func_ov006_020d64c4(void)
-{
-}
