--- src/func_ov006_020d669c.c
@@ -1,4 +1,0 @@
-void func_ov006_020d669c(char (*c)[16]){
-  int i;
-  for(i=0;i<2;i++){ c[i][0x626d]=0; c[i][0x626e]=0; }
-}
