--- src/func_ov006_020d6d7c.c
@@ -1,33 +1,0 @@
-#include "types.h"
-#pragma opt_propagation off
-void func_02012718(void *a, int b);
-
-typedef struct { u8 f0, f1, f2, f3; } Tab;
-extern u8 data_020a0e40;
-extern Tab data_020a0de8[];
-
-void func_ov006_020d6d7c(char *this, int idx) {
-    u8 m = data_020a0e40;
-    int hit = 0;
-    char *b;
-    int dx, dy;
-    if (data_020a0de8[m].f0 != 0) {
-        if (data_020a0de8[m].f1 != 0) hit = 1;
-    }
-    if (hit == 0) return;
-    if (*(u8*)(this + 0x62f6) != 0xff) return;
-    b = this + idx * 0x40;
-    dx = data_020a0de8[m].f2 - (*(int*)(b + 0x4660) >> 0xc);
-    dy = data_020a0de8[m].f3 - (*(int*)(b + 0x4664) >> 0xc);
-    if (dx > 0xc) return;
-    if (dx < -0xc) return;
-    if (dy > 0xf) return;
-    if (dy < -0xf) return;
-    *(u8*)(this + 0x62f6) = *(u8*)(b + 0x4696);
-    *(u8*)(b + 0x4694) = 1;
-    *(u8*)(b + 0x4697) = 2;
-    *(int*)(b + 0x4668) = dx << 0xc;
-    *(int*)(b + 0x466c) = dy << 0xc;
-    *(u8*)(b + 0x469e) = 0;
-    func_02012718((void*)0x1d2, *(int*)(b + 0x4660));
-}
